issue-185: Fixed closestLiveHost method in CassandraConnector
Before that fix, the method used only the hosts which were considered up by the Java driver. However, under a heavy load, this information may not be very up-to-date and in some cases the set of live hosts visible by the Java driver is empty.
The fix comes from the recent changes in LocalNodeFirstLoadBalancing policy - the hosts which are considered down are not filtered out. Instead, they are attached to the tail of the sequence.

Fixes #185