SPARKC-1: Write UDT column values.
To construct a Java driver UDTValue, a reference to UserType (a subclass of DataType) is required. This UserType instance is available from a PreparedStatement, so the code to obtain a TypeConverter for purpose of saving had to be redesigned and moved. Now TypeConverters for saving to Cassandra are no longer associated with ColumnType instances, because ColumnTypes must be Serializable and therefore cannot embed non serializable DataTypes. Instead, a separate method was added to construct a TypeConverter for given DataType and ProtocolVersion.

 Additionally RowWriter interface has been simplified and now it is responsible only for extracting individual column values from a row object, without applying any type conversions to them, binding the PreparedStatement nor estimating the size.  Instead, that  responsibility has been moved to TableWriter so it can be shared between many different RowWriters (and different row representations). This allows TableWriter to create TypeConverters only once before saving multiple rows, which is faster.