SPARKC-94: Move token distance calculation from splitters to TokenFactories.
Further refactoring of TokenFactory. To simplify calculation
of token counts and data sizes in token ranges,
a new concept has been introduced: a ring fraction.

A ring fraction is a number between 0.0 to 1.0
where 1.0 denotes a full ring, while 0.0 would be an empty
token range. Knowing the ring fraction, we can easily
calculate either the number of tokens in the given
token range or, knowing the total amount of data represented
by the ring, the amount of data in the given token range.
This method is very universal and works the same way
for estimating the data size in rows, partitions
or in bytes.

The methods for calculating ring fractions have been added
to TokenFactories.