SPARKC-4: Refactoring: Remove ClassTag from ColumnMapper.
It turns out it is not actually used by most ColumnMapper
implementations, so there is no need to force it on all subclasses.

It was used only by JavaBeanColumnMapper.
Most ColumnMapper implementations use TypeTag, which provides stronger type
information than ClassTag, so keeping both TypeTag and ClassTag would
was redundant. The Java API now obtains a ClassTag
from the Class given as a parameter by the user of the API,
instead of taking it from the ColumnMapper. While this might be
a minor inconvenience for the users of the Java API, because they
need to pass the Class object to some methods previously not
requiring it, this simplifies the code in many places.