SPARKC-147: Modify C* collections from Spark
Previously all statements with collections were treated as overwrites
limiting user's abiilty to modify collections in batch. This patch
provides new options within the ColumnMapper api to allow

- Appending (Lists)
- Prepending (Lists)
- Adding (Sets, Maps, Lists)
- Removing (Lists, Sets)

Removal of elements from maps is currenlty not supported as it would
require much more complicated syntax wrangiling and will be left for a
later ticket if anyone desires the functionality.

To accomplish this a new ColumnRef was created CollectionColumnRef
which contains one additional element CollectionBehavior. TableWriter
now directly takes the ColumnSelector so that it can access these
ColumnBehvaiors when crafting the CQL to do the update.