SPARKC-155: CassandraConnector Host NonLazy Lookup
Instead of using lazy initialization to set the Connector
Host property we will check the cluster for hosts when called.
This is to help avoid situations where other parts of the Connector
rely on this function returning the full list of available hosts
in the Datacenter but have not opened a session yet.