SPARKC-155: Repart uses only DC Local Hosts
This fixes a small bug where the repartition code could
select a replica Endpoint for a given C* key that was not resident in
the same datacenter as the one the connector had chosen. Assuming that
there were no valid endpoints in the datacenter the code would simply
choose a random endpoint in the datacenter as teh preferredLocation.
The new code only chooses from replica endpoints  which are in the same
datacenter as the CassandraConnector connected to.