SPARKC-166: Refactoring: Change Seqs to Sets in PredicatePushDown
PredicatePushDown logic does not depend on the order of predicates
passed to it and duplicates are not allowed.  It divides
the set of predicates into a set of predicates to be pushed down
and a set of all other predicates. This commit makes these properties
explicitly visible in the API.