SPARKC-94: Fetch listen addresses from system.peers table.
Spark RDD getPreferredLocations method must return a list of hostnames
where a partition is best computed at, to enable data-local
computation. The hostnames are given as Strings and must be known to
Spark. Because Spark workers bind to local hostnames, returning a
list of broadcast rpc_addresses may be insufficient, because Spark
may not know them. Therefore, for every rpc_address, we look up
a corresponding Cassandra listen_address in system.peers table.
If Spark nodes bind to the same addresses as Cassandra nodes do,
then the data locality should work.

Limitations: A listen_address of one of the nodes cannot be fetched,
because system.peers table doesn't have data for the local node.
Unfortunately system.local table doesn't store listen_address.
This will be fixed when CASSANDRA-9436 is done.