SPARKC-183: ClosestLiveHost should ignore localhost if down
Previously we could run into a situation where the node the driver was
running on could be down and the application would not run even if a
different node was specified as the intial contact point for the
cluster. This was caused because closestLiveHost was not actually
checking liveness since it relied on sortNodesByProximityAndStatus which
would always place localhost first. This means a dead localhost would be
used before a live neighbor on the driver. This code filters out dead
nodes from the nodes returned by SortNodesByProximityAndStatus
effectively removing the chance of selecting a node which is not
actually up.