SPARKC-183: LNFBalancingPolicy sortByProximityAndStatus to StatusAndProximity
During the review for SPARKC-183 it was decided that we would benefit
from changing the sort method to only put localhost first if it was
alive. This will impact tokenUnawareQueryPlan, so all future queries
which are not token aware will be effected by this change. Effectively
it should not change performance of the Spark Cassandra Connector since
the downed localhost would have been ignored by the driver in the
previous code.