SPARKC-226: Change output.TP_mb_per_sec -> double
We found that having only a whole number granularity made it difficult
to actually tune the output to get maximum throuhgput. This ticket
allows the conf to be set to a double rather than an Int providing the
end user with a much greater ability to control thorughput while
maintaining backwards compatibility.