SPARKC-223: Add withoutLogging for noisy tests
Three tests in particular in the IT suite were constantly logging ERRORs
as the executor processes terminated. These messages were expected since
these tests were meant to raise exceptions. To avoid having noisy logs a
withoutLogging wrapper was added which turns off logging for the
duration of a test case.