SPARKC-271: Let UDTs be converted from GenericRows
When we read a UDT from a C* table it becomes nested generic
rows inside of dataframes. These nested row objects are not convertable
back into UDT types so they cannot be written back into C*. This patch
enables the conversion of nested GenericRows back into UDTS if the
naming is correct.