SPARKC-277: Creates a suite of datatype integration tests
There is a base calss AbstractTypeTest which runs basic insert, read,
write operations for all valid CQL datatypes. Implementing classes
specify both the ScalaReadable type to use and the DriverInsertable type
to use. This should be easily extend able to add additional tests for
new type. We should also focus on extending this coverage to dataframes.