SPARKC-283: Use None for UNSET Parameters in Bound Statements
Protocol Version4 allows a client to not set members of a prepared
statement. This allows a generic prepared statement to be used without
producing tombstones for unset values (previously nulls were inserted).
Within the connector we will now treat None as "UNSET" to allow the end
user to avoid tombstones on saveToCassandra calls.

To accomplish this we change the OptionToNull Converter to
OptionToAnyRef. This now preserves Nones if they exist. The bound
statement builder then passes on any parameters it sees are passed as
None.

To allow backwards compatibility with ProtocolVersion < 4 we need to
special case the BoundStatementBuilder behavior. On versions prior to 4
we still need to place Nulls in any unset bound statement parameters.