SPARKC-287: Upgrade CassandraSqlContext and Catalog to compile v 1.6
I've had to disable some of the functionality around having a seperate
cluster identifier in TableIdentifier. SPARK-10104 changed Table
Identifier from a Seq[String] to a case class with two members Table and
Database. We've raised an issue and will find a workaround in
SPARKC-289. This commit is designed to have Master back up and running.