SPARKC-283: Adds CassandraOption for Skipping Empty Columns
Previously unset columns had to be set to null since the Cassandra
Protocol <= V3 did not support unbound parameters in bound statements.
The newere version of the protocol allows for this, and this patch
enables the connector to take advtange of that behavior.

CassandraOption is created mimicing the Scala Option class itself.
The Option has 3 possible states, Value, Unset and Null. An implicit
conversion is present to transform to a standard scala Option
(Unset and Null -> None). This class can be used as a
default for reading out of Cassandra.

If an Unset is used on a C* version prior to Protcol Version 3 a warning
will be logged that the Unset functionality cannot be used and the value
will be treated as a Delete instead.

Adds test for checking reading value with CassandraOption
Adds tests for all protocol versions in BoundStatementBuilder

Added writeConf parameter ignoreNulls

This parameter will treat all null values as UNSET in bound statement
builder. This can be used when using the DataFrames and SparkSQL apis
which don't allow custom objects.