SPARKC-231: Add Create Table From Dataframe Method
Thanks to @rizzlesauce for the previous commit. This commit adds to the
createTable method for dataframes by allowing the user to set the
partitionkey and clusteringcolumns if they like. In the future we
may want to utilize this functionality within the "write" options of a
data frame so that a user can take advantage of this from the SQL
interface.