SPARKC-323: Fix violation of Partition Contract
The partitions array contains elements with a field "index". This
field must match the actual index in the array or there can be errors
with tasks being infinitely retried. This was broken we implemented
an optimization for task ordering in SPARKC-43. This ticket moves
the "Indexing" of the elements to after the sort applied in SPARKC-43
so the Index field now matches the place of the partition in the
partitions array.