SPARKC-330: Implement org.apache.spark.Partitioner
Implementing org.apache.spark.Partitioner requires determining
the given token of an arbitrary key type of a CassandraTableScanRDD.
CassandraPartitioner has this new responsibilty which it carries out
using a TokenGenerator.

The TokenGenerator is created to determine what the underlying key for a
given Key type would be given a RowWriter.

CassandraRDDPartitioner is renamed CassandraRDDPartitionerGenerator to
avoid confusion with the new class CassandraPartitioner whose sole
responisbility is implementing org.apache.spark.partitioner.

BucketingRangeIndex is created as a Datastructure for fast token
locating. This monotic hashing stucture allows for the quick mapping of
a single element into set of ranges which may or may not span the entier
possible domain.

KeyBy is used as a trigger to apply a partitioner to a given RDD. This
is useful because KeyBy gives us the Type of the key we need to build
a RowWriter for. In the future it would be useful if we could
automatically generate a Partitioner for all K, V rdds but this is
currently impossible.

This change is a bit more dramatic on the Java side as all the keyBy
methods now also require specifying the RowWriterFactory.

Tokens and TokenRanges now have implicit orderings and contain methods
so it is easier to use them in the partitioner when comparing new Tokens
with existing tokenRanges. The driver returns this range (MinToken, Mintoken)
to represent the entire TokenRange. We adjust the RangeBounds Code
to take into account that this special range is actually a WrapAround range.

A public api for applying partitioners is created in
CassandraTableScanRDDFunctions. This adds  implicits applyPartitionerFrom and
applyPartitioner. Both of these methods are KV RDD only so these methods
can be restricted to only to partitioners and other KV RDDs of the
correct types.

These apis are not usable in Java so an additional api keyByAndApply is
added whichwill call both keyBy and then applyPartitionerFrom in
order. This function can be called from a CassandraTableScanJavaRDD
circumventing the inability in Java of calling CassandraRDD methods on
a PairRDD created by keyBy.