SPARKC-347 drop indices that can not be used in predicate push down
We store indices definitions in TableDef in order to use them in

predicate push down. As discussed, currently we do not support predicate
push down for custom and collection indices.

This commit drops these kinds of indices for TableDef#isIndexed and
TableDef#indexedColumns methods. Meaning if an index is not targeted at
single column it never participates in an evaluation of a result for
these two methods.

Nevertheless please note that all indices are present in TableDef.
Unused indices eventually are going to be used in predicate pushdown as
well.

To hide this planned changes from users, some of the TableDef properties
are now private.