SPARKC-360: Disable delayed retrying
Delayed retrying was implemented by Thread.sleep, which was
unfortunately executed by a Netty thread-pool. Blocking that thread pool
may lead to a general lockup of the connector.

Currently, there is no safe way of implementing delayed retries, because
such a feature would have to be implemented in the Java driver.

This change does not introduce API changes. Configuring delayed retrying is
still possible, but will have no effect.