SPARKC-376: Special Case Single Partitions and PartitionKey Restrictions
On older versions of Cassandra 2.1.X, it is illegal to combine Token
restrictions with Partition Key restrictions. To be compatible with
2.1.x single partitions now return a single token range (rather than
multiple with vnodes) and the CassandraTableScanRDD will no append the
TokenRange predicate when the partition key is present. If we did not
use a single TokenRange the query would be applied many times and you
would recieve responses * vnodes.