SPARKC-377 serial hooks execution
Workaround for a problem in which cluster and client connections are
shutdown at the same time, which occasionally leads to deadlock (see
thread dump attached to jira task).

Client and cluster shutting down was executed as JVM shutdownHook, this
commit changes that to sequential invocations (no specific order is
implied).

Also changes back parallelism for Travis to 1.