SPARKC-278: Make Repartition by Cassandra Replica Deterministic
Previously the behavior of repartition by Cassandra Replica was random
within the bounds of assigining keys to the the correct replicas. This
meant that multiple application of the partitioner could result with
rows in different partitions even if they share the same partition key.

The new behvaior pulls the token determination and replica determination
into the partitioner. This removes the need for the
"keyByCassandraReplica" function for repartitionByCassandraReplica.