SPARKC-397: Prevent Null Pointer Exceptions on TTL and WriteTime
Both TTL and WriteTime cannot be set to null and for C* < 2.2 there is
no way to leave the parameter unbound. To avoid null pointers we provide
the following defaults unless modified by the end user.

TTL is set to the table default_ttl
WriteTime is set to currentTimeMillis * 1000L on Insertion

This has required some rather extensive refactoring which has changed
several interfaces. The general thread is pass information via a
ColumnRef down to the BoundStatementBuilder where it can read the
metadata provided and take evasive manuevers when it sees a NULL in a
non-nullable parameter.

WriteOption has been extended so that perRow now also allows for an
additional valueIfNull parameter which can be used to override the
defaults specified above.

Parallel extensions where made to the Java interface to allow for
the same access from java.