SPARKC-387: ColumnType Should be ProtocolVersion Aware
The Connector currently always assumes it is connecting to the newest
version of C*. This leads to some incorrect decisions when checking
against ColumnTypes as PV4 Types are not valid in PV3. To fix this all
of the ColumnType Selection code is now also checks the Protocol Version
and chooses the Correct Type based Protocol Version instead.

In addition, the Integeration tests have been version gated to allow for
accurate testing against previous versions of Cassandra.