SPARKC-355: Add shadedConnector Project
Adds an additional project which mirrors the normal connector build
project except it marks ALL libraries except those we are explictly
shading as provided. This makes it so it's assembly task will only
include the shaded jars we want and no other dependencies.

This assembly task which only includes shaded deps is then set as the
assembly target of spPackages which in coordination with spShade changes
the spDist task to use the shaded assembly jar as it's target rather
than the standard assembly jar with all dependencies.