SPARKC-355: Remove Default Artifact, Publish Only Shaded Jars
Inorder to make sure there is no confusion about what Jar an end user
should be including in their build file, the only artifact that will be
published under the name spark-cassandra-connector will be the shaded
jar. This means all package and packageBin tasks are essentially noops.

publishLocal, and publish will now use the Shaded jar as the the
spark-cassandra-connector artifact.