SPARKC-355: Wire publish and publishLocal to the new package task
With this last change the default connector project is basically a
distribution project. It contains the shaded marked as provided so
pulling from maven will not pull the shaded guava or netty. It's publish
tasks all go through the new packageBin which ends up running the
assembly task from the Shaded project.