SPARKC-419: Fix partitioner on Integer.MIN_VALUE
Tokens might give Integer.MIN_VALUE as hashcode. This will lead to a negative
index into the set of replicas for the corresponding key and fail the task.

The reason is that Math.abs() returns a negative value for Integer.MIN_VALUE
and % gives a negative result in case the dividend is negative.

To fix this the partitioner now applies Math.abs() to the result of the modulo
calculation instead of to the token hash (the dividend). This guarantees that
the index is never negative.

Also adds a test covering ReplicaPartitioner's handling of tokens with hash to
Integer.MIN_VALUE. This test fails without the fix described earlier.