SPARKC-459: Parallelize Travis, Sep Connector Tests, Don't Cache Travis
The Connector tests have been taking a very long time on travis so
it has been switched to run Scala and Cassandra version in parallel. I
also removed caching on travis since it only seems to work on
"containerized" travis which we don't use. I think it was doing strage
things to our dependecy downloading so I removed it for now.

In addtion CassandraConnector specific tests are now run in a seperate
suite since they rely on the Cache not being touched by other tests.