SPARCK-471: Automatic Adjustment of Max Connections
We will boost up the number of CassandraDriver Connections based
on the number of cores allocated to the machine. The user will also be
able to force a parameter for this value. This can be changed at runtime
so we will set it at the creation time of the RDD based on the driver's
information at that time.