SPARCK-479: No parameters with Capital Letters
The Datasource API automatically lower cases all incoming configuration
options. This means any configuration options with Capitals in them
cannot actually be pased to a dataframe. To fix this we will deprecate
all parameters with Capitals and provide automatic translations of them
from their old usage to the new one.

This is done by each "fromSparkConf" method of ours autoamtically
iterating over all DeprecatedConfigParameter objects declared by that
configuration and replacing them if they exist. If we deprecate a
parameter and there is no replacement it will throw an error.

SPARKC-479: 2.0 Merge

Requires deprecating of join_thorughput with the new framework.