SPARKC-475: Add implicit RowWriterFactory for RDD[Row]
Previously when a DataFrame was turned into an RDD by it's `rdd` method
saveToCassandra and joinWithCassandraTable would fail because of the
lack of an implicit RowWriterFactory. To fix this we add an implicit
for a RowWriterFactory for RDD[T <: Row] which ends up mapping to the
SqlRowWriterFactory which we already have written.