SPARKC-470: Allow Writes to Static Columnns and Partition Keys
Previously we were overly strict on checking whether a write was
valid. A Cassandra Table can be written too with only Partition Keys and
Static columns if no clustering columns are included. We ease the check
in TableWriter by only requiring the partition key if all other columns
are Static Columns.