SPARKC-485: Fix ReplicaPartitioner Endpoint Call
Previously endpoints were being calculated with the wrong ByteBuffer.
They needed to use the routing key but were using the Token buffer
instead. To fix this we change the buffer which was passed in to the key
buffer instead.

To fully test this the embedded cassandra class is extended so that it
can create a multi-node cluster. This is then used to make sure or
replica partitioner is accurately dividing nodes by replica and evenly
distributing the keys.