DSP-11539: Add Cassandra Structured Streaming Sink
This allows users to use structured streaming with Cassandra as a Sink.
Previously stream.writeStream.format("o.a.s.s.Cassandra") would have
failed, now we present this interface.

The underlying implementation is very simple, we take the Dataframe as
presented int "addBatch" and use our dataFrame save method to save it.
All options and settings accessible via DataFrames are similarly
applicable for this interface.