DSP-10993: Allow SparkSql to do DirectJoins Against Cassandra
Rebase from Riptano/SCC
Fix Nullability Assignment

Previously all Cassandra fields were assumed nullable. This is incorrect
as partition keys and clustering keys are not nullable. Now nullability
is assigned based on Primary Key status.

Direct Join Implementaiton

Direct Joins are implemented through two classes. The DirectJoinStrategy
and the DirectJoinExec. The Strategy looks through logical plans which
contain a join with a CassandraSource. Those branches are transformed
and rearranged with a DirectJoinExec node in the plan. The replacement
happens as long as the Cassandra Branch times a ratio is larger than the
join key branch. The optimization can also be forced with setting the
direct_join_setting parameter to "on", or "off". The Default "auto"
relies on the size ratio to determine when to make the join.

DseSparkExtensions is used to Load this set of optimizations into the
SparkSession without user intervention.

Fix a Bunch of Scala Warnings

There are a slew of Scala warnings that I cleared up. Most resolved
around PrimaryKeys not being present in match statements.