DSP-8379: Use Cassandra Connector in Hive Metastore
Previously the hive connector used it's own independent
method of connecting to Cassandra. This ended up causing some issues
with the Spark shell shutting down waiting for the metastore's
threadpool to close.

Since the Connector is now inside of DSE we can just access the
CassandraConnector class directly. It will inherit all of the same
options that the SparkConnector would so setting parameters through the
Hadoop configuration is no longer strictly needed for the Metastore.

In the Connector we have added a jWithSessionDo to use withSessionDo
with java Lambdas. I attempted to extend the original method but the
compiler ends up getting confused with some partial functions in Scala.