SPARKC-503: Retry PoolBusy Exceptions, Throttle JWCT Calls
A change in the Cassandra Java driver has caused the executeAsync method
to become fully asynchronous. This means that executeAsync calls can
sometimes return an exception of BusyPoolException, meaning that queries
were issued and no open execution pathway could be found. These
exceptions are always retryable because they indicate the local driver
state is temporarly busy. So we will retry these exceptions.

To make sure we don't have too many retries we will also make sure to
limit the number of max concurrent requests to the current
"requestsPerConnection" as specified in the pooling options. We are
already setting the maxPool size to the number of cores per executor so
this should prohibit having more requests in flight than we have open
connections.