SPARKC-507: Remove excess object creation
Previously there were several places in the read path which
ended up generating objects which were never used or could have been
cached. In particular we change the usage of CassandraRowMetadata which
previously was created for every query. Now only a single metadata will
be made per task in CassandraJoinRDDs.