SCC optimize TimestampParser
Before this change TimestampParser was going throught the list of
parsers and was trying to parse a given string. Once the parsing was
successfull it returned the result. Obviously the further in the list
correct parser is placed, the longer it took to parse the given string.

This commit introduces change in this behaviour. Timestamp parser
chooses a single pattern and tries to parse a given string. It chooses
the pattern based on the given string length.
There are two groups of patterns. The ones that contain 'T' and the ones
that do not. Choosing proper group is based on simple `chartAt` check.

What is important here is the fact that choosing the correct pattern is
done in constant time.

For average scenario (test set contains strings in all available
formats) it shows 15x throughput increase.

Benchmark                    Mode  Cnt       Score      Error  Units

TimestampBenchmark.testNew  thrpt   20  925487.753 ± 4243.462  ops/s
TimestampBenchmark.testOld  thrpt   20   64955.585 ± 1082.897  ops/s

For the worst case scenario (for old parsing - all patterns, except for
the last one, fail to parse) it shows 75 throughput increase.

Benchmark                    Mode  Cnt        Score       Error  Units

TimestampBenchmark.testNew  thrpt   20  3038772.162 ± 39498.833  ops/s
TimestampBenchmark.testOld  thrpt   20    43376.837 ±  1100.900  ops/s