DSP-15193: Increase local connection count in DSEFS and SCC
TPC in DSE 6.0 can't balance the coordination work received from a
single connection, therefore using a single connection to the local node
for the whole executor or for the whole DSEFS can be a serious
bottleneck.

Setting max connections in SCC was not enough, because the driver
is quite reluctant in bumping up the number of connections.

This commit sets the default number of local connections to one less than
the available CPU core count (but not less than 1).

It also adds
new options for fine tuning number of connections in SCC. Now the
numbers for local and remote hosts can be controlled separately.