DSP-15313: Table Options Must be Treated as Case Insensitive
Since Spark 1.3.0 Table Options have been case insenstive internally
even though the signature has not changed in the createRelation api.This
means any time we check a parameter that has been passed in through
TableOptions we must treat the parameter as if it is lower cased.

We also use Locale.ROOT to avoid the "Turkish Lowercase Problem" which
prevents the lower case version of text from being identical to
similarly typed lowercase text.