DSP-15461: Replace all Deprecated SCC Parameter usage
All snake case parameters have been deprecated. So to avoid warnings
we need to change the old parameters to the new camelCased versions.

We used to set size_table_in_mb in spark-defaults but that was only
because we wanted to change the parameter for DSE. Since the connector
is now internal, I have modified the default in that code instead
of in the Spark-Defaults file.