DSP-14523: Add Read Throttling to Spark Connector
Interface Changes:

ReadThroughputMBPS added as an option. None will have
the same behavior as default before, INTEGER.MAX_VALUE mbps.

WriteThroughputMBPS also changed to an option, None having the
previous default behavior.

ReadsPerSecond changed to an option as well.

CP Throttling:

Continuous Paging must be throttled on the server side so we
adjust the ContinousPaging options to match the user defined value.
Since we are given a byte throughput, we use this along with the given
reads per second to set the number of bytes per page per second
option. Absent a Read Throttle the default row based measurements
will be used.

Non-CP Throttling:

The bytes throttled in the incoming read  are counted in the same way
as the inputMetricReader reads byte sizes. This ends up with the size of the
row being calculated twice (once for throttling and once for metrics). We
could extend the Row Class  as a slight optimization in the same way we
extended the Statement Class to add metric information (RichStatement)
but since this is already not our fast path (no-continous
paging) I think just doing a double size calculation is a sufficent solution.

JWCT Throttling:

Previously joinWithCassandraTable could only be throttled on a per
request basis. This implementation allows us to use the input read
throttle on bytes per second to also throttle JWCT calls. Due to our
async nature of performing JWCT our ability to throttle is hampered
so our accuracy in throttling JWCT is expected to not be perfect.