DSP-17044: Dataframe Support for TTL and Writetime
Adds basic mechanical support for reading TTL and WriteTime from the
DataFrame API. To do this we add a set of parameters, provided as
Options to the dataframe.

Reading works like:
  ttl.a=someColumn -> Reads the ttl of a into "someColumn"

Writing can only be done for the whole row:
  ttl=ttlSource -> Uses values from ttlSource as TTL for the whole row

Parsing is done by attempting an Integer parsing first and the falling
back to a rowIdentifier string if a parsing cannot be done.

To allow for reading TTL and WriteTime form DataFrame we add two
expressions one representing each cql function. The expressions are
replaced during the resolution phase of the Catalyst optimization. They
are replaced with new AttributeReferences and the underlying
CassandraSourceRelation is instructed to read into the new reference
using the SCC's implemented TTL and WriteTime instructions.

DSP-17044: Using SparkExtensions to Load Functions at all Times

Before SparkExtensions was run prior to the SessionState being
finalized. This made it impossible to load functions since they need to
be loaded after the session state is built (or while it is built). We
have modified Spark to have an additional extension phase which injects
functions during the SessionState Construction. This removes the need
for loading functions on the SparkShell and automatically will cause
them to be loaded whenever DseSparkExtensions is loaded.

DSP-17044: Add Dataframe writeTime and ttl Helpers

Adds Dataframe helper functions to set TTL and WriteTime. Uses
reflection to determine the currently set output source and only
allows using these helpers when the format is already set to
Cassandra

DSP-17044: Fix WT and TTL for DF Un/Frozen Collections

DSE 6.7 introudced the ability to get TTL from Unfrozen collections.
These are returned as a different type (collection) rather than the
normal Scalar values. To Support this we need to determine whether or
not the the C* table is a frozen collection. We need to determine this
for the functional API as well as when specifying ttl as a conf option.

For functions we need to redo the query plan and change the output type
based on the underlying Cassandra schema