SPARKC-550: Fix Thread Safety Issue in AnyObjectFactory
If multiple threads attempt to convert an object at the same time
there is a chance of racing on filling the temporary arguement buffer.
The buffer is used to add in extra arguments that may be required if the
class has certain external references. To fix this we remove the buffer
entirely and add in the extra args (if they exist) by adding them
to the passed args with sequence concatination.