DSP-17174: Fix Tests broken by SPARK 2.4 Upgrade
We can't just generate Dataframes anymore, instead we must generate new
LocalRelations with the Streaming modifier. Otherwise Spark will
complain when we try to process records.

DSP-17174: Remove unused test

The CheckpoinStreamSpec has not been a working test for several years,
and now it has dependency issues. It seems like the best solution at the
moment is to just clear it out.

DSP-17174: Handle Worker Removed Message in DseAppProxy

This may be the wrong thing to do. I've currently set it to just
Ignore this message when received by our proxy.

DSP-17174: Fix Updating of Local Description and Group Properties

We use SparkContext LocalProperties to propagate the job.description and
groupId of Tinkerpop Olap based Spark Jobs. With Spark 2.4 this stopped
working for DseGraphFrameIntercerptor based tests. I tracked this down
to the Session object that the DGFI was creating somehow persisting the
last used values for these local properties when reused by future
DGFIs. These properties were being correctly set in the SparkContext's
local properties but were being overidden by the old properties hanging
out in the SparkSession. To rememdy this we make a new session for each
DGFI request and remove the local properties from it. This only effects
TP Olap requests which use the DGFI.

DSP-17823: Remove some accidental whitespace