DSP-17593: Make SCC CassandraSourceRelation Spark 2.3 Compatible
In Spark 2.3+ using the "rdd" method on the Dataframe invokes pathways
which are not allowed in Structured Streaming. This means for our source
to be compatible with Spark Streaming we need to work around the direct
"rdd" method. Instead we do what is essentially identical but
circumvents the safeguards preventing users from doing unsafe things.
This method is also used by KafkaSource.