DSP-17697: Reduce the default quiet period to 0 for Spark and tools (master) (#13650)
* DSP-17697: Reduce the default quiet period to 0

Reduce quiet period to 0 by default and make it configurable.
It applies to tools, Cassandra connections in Spark and DSEFS.

* DSP-17697: Fix SearchDseToolCommandsTests

The problem fixed here is that we asserted some operation happened
during reindex but we actually didn't wait for reindexing to complete.
These tests were passing before just by accident - closing Cassandra
driver connection in DseTool introduced artificial sleep of 2 seconds
which allowed these tests to pass. Since we removed that "sleep", these
tests started to fail.

The fix consists of waiting for reindex to complete after it was
requested, before we do assertion. Reindexing is started asynchronously
so we had to use polling wait for this purpose.

* Address review comments