DSP-17823: Fix CassandraHiveMetastore ListPartitions Method
Previously when presented with a partial partion spec the
listPartitionsPsWithAuth method would always return empty unless
the partition values were stored as is in the part_vals (mislabled as
maxParts). Unfortunately this was not accurate as the part_vals is
defined as an array of values in the same order as the partition keys
with empty strings as placeholders when no partial spec was present.
Spark 2.4 started using partial specs (empty by default)
and caused hive stored tables to no longer correctly report their
partition information.

To fix this we use a modified version of the name matching algorithm in
the CachedStore from Apache Hive. This code generates a matching regex
like key=.*/key2=value and matches this against the partition name which
is represented like key=1/key2=value.